+++ app/src/main/java/com/imusic/fragment/child/my_music/artists/details/ArtistDetailViewModel.java
@@ -17,7 +17,7 @@ public class ArtistDetailViewModel extends AndroidViewModel {
         mRepository = new ArtistDetailRepository(application);
     }
 
-    LiveData<List<Song>> getListLiveDataSongArtist(int artistId) {
+    LiveData<List<Song>> getListLiveDataSongArtist(long artistId) {
         return mRepository.getListLiveDataArtistSong(artistId);
     }
 }
